knx Bus Simulator
This is a tool for simulation of a knx bus on Mac OS.
Supported environments are Mac OS and Rasperry PI. Compiled on Mac OS
this is a plain simulation. Compiled on Raspberry PI it hooks up
through a serial device via a TPUART to a real KNX Bus.
PLEASE NOTE:
Nothing contained in this package, which is provided as is and with the
purpose of PLAIN SIMULATION, has been run through any of the KNX
procedures for certification!!!!!
Thx to Teunis van Beelen, who’s serial library is used to hook up the
real serial device.
For now, you will have to dig through the source code to figure out how
this system works. A detailed documentation will be provided soon.
(right now I’m too busy with my construction project, including this
bus integration).